requirements.txt
geopy>=2.3.0
imageio>=2.31.1
matplotlib>=3.7.2
numpy>=1.25.1
Pillow==9.0.0
Requests>=2.31.0
scikit_learn>=1.3.0
scikit_image
staticmap>=0.5.5
torch>=2.0.1
torchmetrics>=1.0.1
torchvision>=0.15.2
tqdm>=4.65.0
urllib3>=2.0.3
utm>=0.7.0
